Changed name of module to Forked from ForkedResource.
Having module and class be same name can cause extra work to distinguish stuff. Eg. ForkedResource.Error was giving an error because it was accessing the class, not the module. You can work around, but easier to just have a different name.

+++ Sources/Forked/ForkedResource+Updating.swift
@@ -14,6 +14,7 @@ public extension ForkedResource {
     
     func delete(_ fork: Fork) throws {
         try serialize {
+            guard fork != .main else { throw Error.attemptToDeleteMainFork }
             try repository.delete(fork)
         }
     }
